+++ src/Powercord/plugins/pc-moduleManager/util/index.js
@@ -1,7 +1,9 @@
 const getRepoInfo = require('./getInfo');
 const cloneRepo = require('./cloneRepo');
+const { resp } = require('./resp');
 
 module.exports = {
   getRepoInfo,
-  cloneRepo
+  cloneRepo,
+  resp
 };
